+++ Tests/Functional/AppKernel.php
@@ -16,7 +16,7 @@ class AppKernel extends Kernel
 
         $fs = new Filesystem();
         if (!$fs->isAbsolutePath($config)) {
-            $config = __DIR__.'/config/'.$config;
+            $config = __DIR__.'/TestBundle/Resources/config/'.$config;
         }
 
         if (!file_exists($config)) {