+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
     - Added missing `$persist` parameter to `set` method
 - `JMS\Payment\CoreBundle\EntityExtendedDataType::convertToDatabaseValue` now throws an exception when attempting to convert an object which does not implement `JMS\Payment\CoreBundle\Model\ExtendedDataInterface`.
 
+### Deprecated
+- The service `payment.encryption_service` has been deprecated and is now an alias to `payment.crypto.mcrypt`. Parameters specified for `payment.encryption_service` are automatically set for `payment.crypto.mcrypt` so no changes are required in service configuration until `payment.encryption_service` is removed in 2.0.
+
 ## [1.2.0] - 2016-10-03
 ### Added
 - Added support for Symfony 3.0. Note that Symfony 3.0 introduces BC breaks. This means that you'll probably need to do more than simply updating to version `1.2.0` of this bundle for your code to keep working under Symfony 3.0. Please see Symfony's [Upgrade Guide](https://github.com/symfony/symfony/blob/master/UPGRADE-3.0.md) for information on what you need to change.